move boxstats out of numeric header
